+++ share/src/model/assemble_reqres.rs
@@ -30,7 +30,7 @@ pub struct AcAsmListReq {
     #[serde_as(as = "NoneAsEmptyString")]
     pub prev_id: Option<i32>,
     #[serde_as(as = "NoneAsEmptyString")]
-    pub size: Option<i64>,
+    pub size: Option<i32>,
 }
 
 impl Default for AcAsmListReq {
